fix(LT-4278): remove parallel connection checking
Removed parallel connection checking to avoid multiple unknown variables while investigating connection issues, raising TransportException immediately cause we can'r proceed if a single queue was not subscribed on. It is potentially dangerous and data lossy to start trading core with just a part of publishers or subscribers. Additionally, this way exception raising saves the stack trace.